Create BusFabric tests for the reading and writing
- misaligned reads
- atomic xor, set, clear writes
- narrow writes with replications and zeros for unused lines

+++ Venture/Debug/GdbConnectionHandler.cs
@@ -269,7 +269,7 @@ public class GdbConnectionHandler : ConnectionHandler
         {
             try
             {
-                int registerCount = emulator.Registers.Length;
+                int registerCount = (int)emulator.Registers.Length;
                 int hexCharsPerReg = 8;
                 int totalLength = registerCount * hexCharsPerReg;
